requirements.txt
asgiref==3.8.1
blinker==1.9.0
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudevents==1.11.0
cryptography==44.0.0
Deprecated==1.2.15
deprecation==2.1.0
Django==5.1.4
djangorestframework==3.15.2
firebase-admin==6.6.0
firestore==0.0.8
Flask==3.1.0
functions-framework==3.8.2
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-pubsub==2.27.1
google-cloud-retail==1.24.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msgpack==1.1.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==24.2
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.0
requests==2.32.3
rsa==4.9
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.0
zipp==3.21.0
